fix: 语义复验与收尾清理
- chore: 固定开发环境的玩家名（runClient 用 dev_username）
- fix: 修掉语义核对发现的 3 处与上游不一致
- fix: AutoThrow 的附近药水过滤收窄到喷溅药水
- chore: 清理本次同步引入的 63 条未使用导入

+++ src/main/java/me/matl114/hacks/modules/combat/ElytraBot.java
@@ -44,7 +44,6 @@ import me.matl114.utils.*;
 import me.matl114.utils.algorithms.StateMachine;
 import me.matl114.utils.entity.PlayerInputUtils;
 import me.matl114.versioned.api.VDataFlag;
-import me.matl114.versioned.api.VDrawContext;
 import me.matl114.versioned.api.VItem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.core.BlockPos;
@@ -1787,7 +1786,7 @@ public class ElytraBot extends BaseModule {
                         Vec3 predictionPosition = PositionPredict.INSTANCE
                                 .attackPredictArgument
                                 .get()
-                                .predictWithExtraTicks(base.target, (int) predictionTicks);
+                                .predictWithExtraTicks(base.target, predictionTicks);
                         Vec3 untrustedDirection = calculateTargetDirection(predictionPosition);
                         if (untrustedDirection.y < 0 && conditionMovement.test(untrustedDirection)) {
                             movementDirection = untrustedDirection;
